feat: add smoke tests for all 4 languages + fix Go int64/int32/time.Duration type assertions
Adds a 4th defense layer (smoke test) to the CI pipeline:
- Go: tests/ with errReader that verifies 3-tier fallback returns defaults
- Java: SmokeTest.java with null reader getter validation
- Python: smoke_test.py with None reader method calls
- C: smoke_test.c with dummy vtable fallback verification

Fixes Go codegen backend bug where hardcoded defaults in map[string]interface{} lacked
explicit type wrappers for int64/int32/time.Duration — untyped numeric literals default
to 'int' in Go, causing type assertion panics in adapter methods.

+++ biz-adapters/go/douyin-mall/adapters/blessstar/review_adapter.go
@@ -30,8 +30,8 @@ func NewReviewConfigAdapter(reader ports.ConfigReader) ports.ReviewConfig {
 	return &ReviewConfigAdapter{
 		reader: reader,
 		hardcodedDefaults: map[string]interface{}{
-			"RatingMin": 1,
-			"RatingMax": 5,
+			"RatingMin": int32(1),
+			"RatingMax": int32(5),
 		},
 	}
 }
